+++ CHANGELOG.md
@@ -1,7 +1,20 @@
 ## Azure NetAppFiles SDK Sample for .NETCore Changelog
 
+* [1.2.0 (2021-09-07)](#1.2.0 (2021-09-07))
 * [1.1.0 (2019-10-22)](#1.1.0 (2019-10-22))
 * [1.0.0 (2019-07-29)](#1.0.0 (2019-07-29))
+
+# 1.2.0 (2021-09-07)
+*Features*
+* SDKs updated to latest version with some minor code changes
+* Upgraded target framework to NET 5.0 
+* Clean up resources logic were added
+
+*Bug Fixes*
+* Support new protocol type NFSv4.1 
+
+*Breaking Changes*
+* N/A
 
 # 1.1.0 (2019-10-22)
 *Features*